feat: add whatsapp broadcast message api and delivery order message

+++ use_case/product_use_case.go
@@ -381,7 +381,9 @@ func (u *productUseCase) Update(ctx context.Context, request dto_request.Product
 		}),
 	)
 
-	u.baseFileUseCase.mainFilesystem.Delete(toBeDeletedImageFile.Path)
+	if toBeDeletedImageFile != nil {
+		u.baseFileUseCase.mainFilesystem.Delete(toBeDeletedImageFile.Path)
+	}
 
 	u.mustLoadProductDatas(ctx, []*model.Product{&product}, productLoaderParams{
 		productStock: true,
